requirements.txt
Flask==3.0.0
google-cloud-firestore==2.13.1
google-auth==2.27.0
requests==2.31.0
